Safari-611.1.10/LayoutTests/http/tests/misc/gamepads-insecure-expected.txt

CONSOLE MESSAGE: Navigator.getGamepads() will be removed from insecure contexts in a future release
Test that accessing navigator.getGamepads from an insecure context logs a warning
